Update index.js
Hey theusaf! I noticed that the LATESTCOMIC template returns, for example, " 2819" instead of "2819" (with a leading space). I'm not sure if the leading space is required, but if it's not, could it be removed? It messes the "List of comics" template and I think it's unintuitive: one would expect a template like that to return a single number. Thanks!